SV fixed
adjusted the sql so it is only for GS style victory. (We balance assuming GS anyway)
Adjusted the lua so it is now a simple passthrough
Adjusted trhe SV style based on Chuck's feedback: so the distance to travel is bigger and the time is speed up at tech level now.